back/requirements.txt
# Core
Flask==3.0.0
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# CORS
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# OCR (for CarInfo)
pytesseract==0.3.10
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Logging
python-json-logger==2.0.7

# Security
cryptography==42.0.0

# Development
flask-debugtoolbar==0.14.1